+++ packages/lib-common/src/scopeSupportFacets/languageScopeSupport.ts
@@ -27,6 +27,7 @@ import { scalaScopeSupport } from "./scala";
 import { scmScopeSupport } from "./scm";
 import type { LanguageScopeSupportFacetMap } from "./scopeSupportFacets.types";
 import { scssScopeSupport } from "./scss";
+import { shellscriptScopeSupport } from "./shellscript";
 import { swiftScopeSupport } from "./swift";
 import { talonScopeSupport } from "./talon";
 import { talonListScopeSupport } from "./talonList";
@@ -69,6 +70,7 @@ export const languageScopeSupport: Record<
   scala: scalaScopeSupport,
   scm: scmScopeSupport,
   scss: scssScopeSupport,
+  shellscript: shellscriptScopeSupport,
   swift: swiftScopeSupport,
   "talon-list": talonListScopeSupport,
   talon: talonScopeSupport,